Test chop-up:
- moved default tests to new class
- moved exceptions under exceptions package
- added new test classes to TestSuite
