+++ RPGManager_development/src/rpg_database/character_sheet/SpecializationClasses.java
@@ -1,5 +1,7 @@
 package rpg_database.character_sheet;
 
+import rpg_database.character_sheet.exceptions.InvalidCharacterClassException;
+
 public enum SpecializationClasses implements CustomSetter<SpecializationClasses> {
 	NOT_APPLICABLE("N/A"), ARCANE_WARRIOR("Arcane Warrior", BaseClasses.MAGE), ASSASSIN("Assassin", BaseClasses.ROGUE),
 	BARD("Bard", BaseClasses.ROGUE), BERSERKER("Berserker", BaseClasses.WARRIOR), BLOOD_MAGE("Blood Mage", BaseClasses.MAGE),
