+++ RPGManager_development/src/rpg_database/character_sheet/exceptions/InvalidCharacterClassException.java
@@ -1,4 +1,4 @@
-package rpg_database.character_sheet;
+package rpg_database.character_sheet.exceptions;
 
 public class InvalidCharacterClassException extends RuntimeException {
 	public InvalidCharacterClassException(String message) {
