+++ RPGManager_development/test/unit_test/character_sheet_unit_tests/CharacterSheetUnitTests.java
@@ -8,25 +8,127 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import rpg_database.character_sheet.Background;
-import rpg_database.character_sheet.BaseClasses;
 import rpg_database.character_sheet.CharacterSheet;
 import rpg_database.character_sheet.Fields;
 import rpg_database.character_sheet.Gender;
-import rpg_database.character_sheet.InvalidCharacterClassException;
-import rpg_database.character_sheet.Money;
-import rpg_database.character_sheet.SpecializationClasses;
 
 public class CharacterSheetUnitTests {
 	// fields
-
-	final CharacterSheet defaultCharacterSheet = new CharacterSheet("DefultCharacterSheet");
 	CharacterSheet testCharacterSheet = new CharacterSheet("TestCharacterSheet");
 
 	// test methods
 	@Rule
 	public ExpectedException thrown = ExpectedException.none();
 
+	// exception tests
+	@Test
+	public void expectException_SetCharacterNameMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
+		final double malformedInput = 26.12;
+		testCharacterSheet.setData(Fields.NAME, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterAgeMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
+		final String malformedInput = "MALFORMED INPUT";
+		testCharacterSheet.setData(Fields.AGE, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterGenderMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class,
+				"class java.lang.Double value is not an instance of class rpg_database.character_sheet.Gender");
+		final double malformedInput = 26.12;
+		testCharacterSheet.setData(Fields.GENDER, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterXPMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
+		final String malformedInput = "MALFORMED INPUT";
+		testCharacterSheet.setData(Fields.XP, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterSpeedMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
+		final String malformedInput = "MALFORMED INPUT";
+		testCharacterSheet.setData(Fields.SPEED, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterLevelMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
+		final String malformedInput = "MALFORMED INPUT";
+		testCharacterSheet.setData(Fields.LEVEL, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterDefenseMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
+		final String malformedInput = "MALFORMED INPUT";
+		testCharacterSheet.setData(Fields.DEFENSE, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterArmorMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
+		final String malformedInput = "MALFORMED INPUT";
+		testCharacterSheet.setData(Fields.ARMOR_RATING, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterHealthMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
+		final String malformedInput = "MALFORMED INPUT";
+		testCharacterSheet.setData(Fields.HEALTH_POINTS, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterManaMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
+		final String malformedInput = "MALFORMED INPUT";
+		testCharacterSheet.setData(Fields.MANA_POINTS, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterAppearanceMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
+		final double malformedInput = 21.10;
+		testCharacterSheet.setData(Fields.APPEARANCE, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterDistinguishingFeaturesMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
+		final double malformedInput = 21.10;
+		testCharacterSheet.setData(Fields.DISTINGUISHING_FEATURES, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterOftenUsedEquipmentMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
+		final double malformedInput = 21.10;
+		testCharacterSheet.setData(Fields.OFTEN_USED_EQUIPMENT, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterGoalsAndTiesMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
+		final double malformedInput = 21.10;
+		testCharacterSheet.setData(Fields.GOALS_AND_TIES, malformedInput);
+	}
+
+	@Test
+	public void expectException_SetCharacterEquipmentMalformedInput() {
+		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
+		final double malformedInput = 21.10;
+		testCharacterSheet.setData(Fields.EQUIPMENT, malformedInput);
+	}
+	// functional unit tests
+
+	// TODO! Character exceptions? special characters? @&!%? system unallowed
 	@Test
 	public void testGetCharacterSheetEntryName_Tibor() {
 		final String expectedCharacterName = "Tibor_karilapja";
@@ -46,38 +148,12 @@ public class CharacterSheetUnitTests {
 	}
 
 	@Test
-	public void expectException_SetCharacterNameMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
-		final double malformedInput = 26.12;
-		testCharacterSheet.setData(Fields.NAME, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterName() {
-		String characterName = (String) defaultCharacterSheet.getData(Fields.NAME);
-		assertEquals("", characterName);
-	}
-
-	@Test
 	public void testSetCharacterNameTibor() {
 		final String expectedCharacterName = "Tibor";
 
 		testCharacterSheet.setData(Fields.NAME, expectedCharacterName);
 		String characterName = (String) testCharacterSheet.getData(Fields.NAME);
 		assertEquals(expectedCharacterName, characterName);
-	}
-
-	@Test
-	public void expectException_SetCharacterAgeMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.AGE, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterAge() {
-		int age = defaultCharacterSheet.getData(Fields.AGE);
-		assertEquals(0, age);
 	}
 
 	@Test
@@ -90,37 +166,10 @@ public class CharacterSheetUnitTests {
 	}
 
 	@Test
-	public void expectException_SetCharacterGenderMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class,
-				"class java.lang.Double value is not an instance of class rpg_database.character_sheet.Gender");
-		final double malformedInput = 26.12;
-		testCharacterSheet.setData(Fields.GENDER, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterGender() {
-		Gender gender = defaultCharacterSheet.getData(Fields.GENDER);
-		assertEquals(Gender.MALE, gender);
-	}
-
-	@Test
 	public void testSetCharacterGenderFemale() {
 		testCharacterSheet.setData(Fields.GENDER, Gender.FEMALE);
 		Gender gender = testCharacterSheet.getData(Fields.GENDER);
 		assertEquals(Gender.FEMALE, gender);
-	}
-
-	@Test
-	public void expectException_SetCharacterXPMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.XP, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterExperience() {
-		int xp = defaultCharacterSheet.getData(Fields.XP);
-		assertEquals(0, xp);
 	}
 
 	@Test
@@ -138,71 +187,10 @@ public class CharacterSheetUnitTests {
 	}
 
 	@Test
-	public void expectException_SetCharacterSpeedMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.SPEED, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterSpeed() {
-		int speed = defaultCharacterSheet.getData(Fields.SPEED);
-		assertEquals(0, speed);
-	}
-
-	@Test
 	public void testSetCharacterSpeed12() {
 		testCharacterSheet.setData(Fields.SPEED, 12);
 		int speed = testCharacterSheet.getData(Fields.SPEED);
 		assertEquals(12, speed);
-	}
-
-	@Test
-	public void expectException_SetCharacterMoneyMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.MONEY, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterMoney() {
-		int money = defaultCharacterSheet.getData(Fields.MONEY);
-		assertEquals(0, money);
-	}
-
-	@Test
-	public void testSetCharacterMoney1200() {
-		testCharacterSheet.setData(Fields.MONEY, 1200);
-		int money = testCharacterSheet.getData(Fields.MONEY);
-		assertEquals(1200, money);
-	}
-
-	@Test
-	public void testGetDefaultCharacterMoneyCoins() {
-		int money = defaultCharacterSheet.getData(Fields.MONEY);
-		Money coins = new Money(money);
-		assertEquals("You have 0 gold, 0 silver and 0 copper coins", coins.toWriteConsole());
-	}
-
-	@Test
-	public void testSetCharacterMoneyCoins124576() {
-		testCharacterSheet.setData(Fields.MONEY, 124576);
-		int money = testCharacterSheet.getData(Fields.MONEY);
-		Money coins = new Money(money);
-		assertEquals("You have 12 gold, 45 silver and 76 copper coins", coins.toWriteConsole());
-	}
-
-	@Test
-	public void expectException_SetCharacterLevelMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.LEVEL, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterLevel() {
-		int level = defaultCharacterSheet.getData(Fields.LEVEL);
-		assertEquals(0, level);
 	}
 
 	@Test
@@ -213,36 +201,10 @@ public class CharacterSheetUnitTests {
 	}
 
 	@Test
-	public void expectException_SetCharacterDefenseMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.DEFENSE, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterDefense() {
-		int defense = defaultCharacterSheet.getData(Fields.DEFENSE);
-		assertEquals(0, defense);
-	}
-
-	@Test
 	public void testSetCharacterDefense15() {
 		testCharacterSheet.setData(Fields.DEFENSE, 15);
 		int defense = testCharacterSheet.getData(Fields.DEFENSE);
 		assertEquals(15, defense);
-	}
-
-	@Test
-	public void expectException_SetCharacterArmorMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.ARMOR_RATING, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterArmor() {
-		int armor = defaultCharacterSheet.getData(Fields.ARMOR_RATING);
-		assertEquals(0, armor);
 	}
 
 	@Test
@@ -253,36 +215,10 @@ public class CharacterSheetUnitTests {
 	}
 
 	@Test
-	public void expectException_SetCharacterHealthMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.HEALTH_POINTS, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterHealth() {
-		int health = defaultCharacterSheet.getData(Fields.HEALTH_POINTS);
-		assertEquals(0, health);
-	}
-
-	@Test
 	public void testSetCharacterHealth80() {
 		testCharacterSheet.setData(Fields.HEALTH_POINTS, 80);
 		int health = testCharacterSheet.getData(Fields.HEALTH_POINTS);
 		assertEquals(80, health);
-	}
-
-	@Test
-	public void expectException_SetCharacterManaMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.String value is not an instance of class java.lang.Integer");
-		final String malformedInput = "MALFORMED INPUT";
-		testCharacterSheet.setData(Fields.MANA_POINTS, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterMana() {
-		int mana = defaultCharacterSheet.getData(Fields.MANA_POINTS);
-		assertEquals(0, mana);
 	}
 
 	@Test
@@ -293,36 +229,10 @@ public class CharacterSheetUnitTests {
 	}
 
 	@Test
-	public void expectException_SetCharacterAppearanceMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
-		final double malformedInput = 21.10;
-		testCharacterSheet.setData(Fields.APPEARANCE, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterAppearance() {
-		String appearance = defaultCharacterSheet.getData(Fields.APPEARANCE);
-		assertEquals("", appearance);
-	}
-
-	@Test
 	public void testSetCharacterAppearanceHandsome() {
 		testCharacterSheet.setData(Fields.APPEARANCE, "Handsome");
 		String appearance = testCharacterSheet.getData(Fields.APPEARANCE);
 		assertEquals("Handsome", appearance);
-	}
-
-	@Test
-	public void expectException_SetCharacterDistinguishingFeaturesMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
-		final double malformedInput = 21.10;
-		testCharacterSheet.setData(Fields.DISTINGUISHING_FEATURES, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterDistinguishingFeatures() {
-		String distinguishingFeatures = defaultCharacterSheet.getData(Fields.DISTINGUISHING_FEATURES);
-		assertEquals("", distinguishingFeatures);
 	}
 
 	@Test
@@ -333,36 +243,10 @@ public class CharacterSheetUnitTests {
 	}
 
 	@Test
-	public void expectException_SetCharacterOftenUsedEquipmentMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
-		final double malformedInput = 21.10;
-		testCharacterSheet.setData(Fields.OFTEN_USED_EQUIPMENT, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterOftenUsedEquipment() {
-		String oftenUsedEquipment = defaultCharacterSheet.getData(Fields.OFTEN_USED_EQUIPMENT);
-		assertEquals("", oftenUsedEquipment);
-	}
-
-	@Test
 	public void testSetCharacterOftenUsedEquipmentPipe() {
 		testCharacterSheet.setData(Fields.OFTEN_USED_EQUIPMENT, "Pipe");
 		String oftenUsedEquipment = testCharacterSheet.getData(Fields.OFTEN_USED_EQUIPMENT);
 		assertEquals("Pipe", oftenUsedEquipment);
-	}
-
-	@Test
-	public void expectException_SetCharacterGoalsAndTiesMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
-		final double malformedInput = 21.10;
-		testCharacterSheet.setData(Fields.GOALS_AND_TIES, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterGoalsAndTies() {
-		String goalsAndTies = defaultCharacterSheet.getData(Fields.GOALS_AND_TIES);
-		assertEquals("", goalsAndTies);
 	}
 
 	@Test
@@ -373,19 +257,6 @@ public class CharacterSheetUnitTests {
 	}
 
 	@Test
-	public void expectException_SetCharacterEquipmentMalformedInput() {
-		expectExceptionWithMessage(InvalidParameterException.class, "class java.lang.Double value is not an instance of class java.lang.String");
-		final double malformedInput = 21.10;
-		testCharacterSheet.setData(Fields.EQUIPMENT, malformedInput);
-	}
-
-	@Test
-	public void testGetDefaultCharacterEquipment() {
-		String equipment = defaultCharacterSheet.getData(Fields.EQUIPMENT);
-		assertEquals("", equipment);
-	}
-
-	@Test
 	public void testSetCharacterEquipmentHasABackpack() {
 		testCharacterSheet.setData(Fields.EQUIPMENT, "HasABackpack");
 		String equipment = testCharacterSheet.getData(Fields.EQUIPMENT);
@@ -393,7 +264,6 @@ public class CharacterSheetUnitTests {
 	}
 
 	// private methods
-
 	private void expectExceptionWithMessage(Class<? extends Exception> exceptionClass, String message) {
 		thrown.expect(exceptionClass);
 		thrown.expectMessage(message);
