+++ RPGManager_development/test/unit_test/character_sheet_unit_tests/InvalidClassAndSpecializationPairingsTests.java
@@ -14,8 +14,8 @@ import org.junit.runners.Parameterized.Parameters;
 import rpg_database.character_sheet.BaseClasses;
 import rpg_database.character_sheet.CharacterSheet;
 import rpg_database.character_sheet.Fields;
-import rpg_database.character_sheet.InvalidCharacterClassException;
 import rpg_database.character_sheet.SpecializationClasses;
+import rpg_database.character_sheet.exceptions.InvalidCharacterClassException;
 import unit_test.character_sheet_unit_tests.resources.SpecializationCompatibilityData;
 
 @RunWith(Parameterized.class)
